sm6150-common: Build soundtrigger 2.0 and 2.3
libc    : Unable to set property ctl.interface_start to android.hardware.soundtrigger@2.0::ISoundTriggerHw/default: error code: 0x20
hwservicemanager: Tried to start android.hardware.soundtrigger@2.0::ISoundTriggerHw/default as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.
I hwservicemanager: getTransport: Cannot find entry android.hardware.soundtrigger@2.3::ISoundTriggerHw/default in either framework or device VINTF manifest.